Move NPCStats class and related types to Tables module
